+++ main.go
@@ -17,7 +17,7 @@ const (
 	DefaultPfRegisterPath           = "api/v1/node/register"
 	DefaultPfStoreMetricsPath       = "api/v1/node/store_metrics"
 	DefaultPfListContainersPath     = "api/v1/node/containers/scheduled"
-	DefaultPfUpdateIpaddress        = "api/v1/node/containers/ipaddress"
+	DefaultPfUpdateIpaddressPath    = "api/v1/node/containers/ipaddress"
 	DefaultPfMarkProvisionedPath    = "api/v1/node/containers/mark_provisioned"
 	DefaultPfMarkProvisionErrorPath = "api/v1/node/containers/mark_provision_error"
 	DefaultPfMarkDeletedPath        = "api/v1/node/containers/mark_deleted"
@@ -29,7 +29,7 @@ const (
 	EnvPfRegisterPath           = "PF_REGISTER_PATH"
 	EnvPfStoreMetricsPath       = "PF_STORE_METRICS_PATH"
 	EnvPfListContainersPath     = "PF_LIST_CONTAINERS_PATH"
-	EnvPfUpdateIpaddress        = "PF_UPDATE_IPADDRESS"
+	EnvPfUpdateIpaddressPath    = "PF_UPDATE_IPADDRESS_PATH"
 	EnvPfMarkProvisionedPath    = "PF_MARK_PROVISIONED_PATH"
 	EnvPfMarkProvisionErrorPath = "PF_MARK_PROVISION_ERROR_PATH"
 	EnvPfMarkDeletedPath        = "PF_MARK_DELETED_PATH"
@@ -57,7 +57,7 @@ func init() {
 	PfApiPath["Register"], _ = envkit.GetString(EnvPfRegisterPath, DefaultPfRegisterPath)
 	PfApiPath["StoreMetrics"], _ = envkit.GetString(EnvPfStoreMetricsPath, DefaultPfStoreMetricsPath)
 	PfApiPath["ListContainers"], _ = envkit.GetString(EnvPfListContainersPath, DefaultPfListContainersPath)
-	PfApiPath["UpdateIpaddress"], _ = envkit.GetString(EnvPfUpdateIpaddress, DefaultPfUpdateIpaddress)
+	PfApiPath["UpdateIpaddress"], _ = envkit.GetString(EnvPfUpdateIpaddressPath, DefaultPfUpdateIpaddressPath)
 	PfApiPath["MarkProvisioned"], _ = envkit.GetString(EnvPfMarkProvisionedPath, DefaultPfMarkProvisionedPath)
 	PfApiPath["MarkProvisionError"], _ = envkit.GetString(EnvPfMarkProvisionErrorPath, DefaultPfMarkProvisionErrorPath)
 	PfApiPath["MarkDeleted"], _ = envkit.GetString(EnvPfMarkDeletedPath, DefaultPfMarkDeletedPath)