+++ metrics/metrics.go
@@ -1,15 +1,15 @@
 package metrics
 
 import (
-	"github.com/pathfinder-cm/pathfinder-agent/model"
 	"github.com/pathfinder-cm/pathfinder-agent/util"
+	"github.com/pathfinder-cm/pathfinder-go-client/pfmodel"
 	"github.com/shirou/gopsutil/mem"
 	log "github.com/sirupsen/logrus"
 )
 
 type Metrics interface {
-	Collect() model.Metrics
-	collectMemoryMetrics() model.Memory
+	Collect() pfmodel.Metrics
+	collectMemoryMetrics() pfmodel.Memory
 }
 
 type metrics struct{}
@@ -18,21 +18,21 @@ func NewMetrics() Metrics {
 	return &metrics{}
 }
 
-func (m *metrics) Collect() model.Metrics {
+func (m *metrics) Collect() pfmodel.Metrics {
 	memory := m.collectMemoryMetrics()
 
-	return model.Metrics{
+	return pfmodel.Metrics{
 		Memory: memory,
 	}
 }
 
-func (m *metrics) collectMemoryMetrics() model.Memory {
+func (m *metrics) collectMemoryMetrics() pfmodel.Memory {
 	vmStat, err := mem.VirtualMemory()
 	if err != nil {
 		log.Error(err.Error())
 	}
 
-	return model.Memory{
+	return pfmodel.Memory{
 		Used:  util.BToMb(vmStat.Used),
 		Free:  util.BToMb(vmStat.Free),
 		Total: util.BToMb(vmStat.Total),