+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changelog
 
 #### 0.3.0
-- Add feature to store metrics from node(s) such as memory info (only this for now)
+- Add feature to collect metrics from node and store it on server
 
 #### 0.2.0
 - Model and pfclient are extracted to [pathfinder-go-client](https://github.com/pathfinder-cm/pathfinder-go-client) repo